// GradeSession.Client/src/shared/api/generated/model/sessionParticipantRole.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */

export type SessionParticipantRole = typeof SessionParticipantRole[keyof typeof SessionParticipantRole];


export const SessionParticipantRole = {
  Secretary: 'Secretary',
  CommitteeMember: 'CommitteeMember',
} as const;