// GradeSession.Client/src/shared/api/generated/model/sessionStatus.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  Draft: 'Draft',
  Active: 'Active',
  Closed: 'Closed',
} as const;