// GradeSession.Client/src/shared/api/generated/model/committeeSlotProgressResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */

export interface CommitteeSlotProgressResponse {
  /** @nullable */
  committeeSlotId?: string | null;
  /** @nullable */
  claimedByUserId?: string | null;
  totalStudents?: number;
  studentsWithAnyAssessment?: number;
  completedStudents?: number;
  totalCriterionScores?: number;
  filledCriterionScores?: number;
  totalFinalGrades?: number;
  filledFinalGrades?: number;
}