// GradeSession.Client/src/shared/api/generated/model/committeeSlotResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */
import type { SessionParticipantRole } from './sessionParticipantRole';

export interface CommitteeSlotResponse {
  /** @nullable */
  id?: string | null;
  role?: SessionParticipantRole;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  shortName?: string | null;
  /** @nullable */
  claimedByUserId?: string | null;
  /** @nullable */
  claimedAt?: string | null;
}