// GradeSession.Client/src/shared/api/generated/model/sessionProgressResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */
import type { CommitteeSlotProgressResponse } from './committeeSlotProgressResponse';

export interface SessionProgressResponse {
  /** @nullable */
  sessionId?: string | null;
  studentsCount?: number;
  criteriaCount?: number;
  /** @nullable */
  committee?: CommitteeSlotProgressResponse[] | null;
}