// GradeSession.Client/src/shared/api/generated/model/sessionResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */
import type { CommitteeSlotResponse } from './committeeSlotResponse';
import type { CriterionResponse } from './criterionResponse';
import type { SessionSettingsResponse } from './sessionSettingsResponse';
import type { SessionStatus } from './sessionStatus';
import type { StudentResponse } from './studentResponse';

export interface SessionResponse {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  createdByUserId?: string | null;
  status?: SessionStatus;
  settings?: SessionSettingsResponse;
  createdAt?: string;
  updatedAt?: string;
  /** @nullable */
  students?: StudentResponse[] | null;
  /** @nullable */
  criteria?: CriterionResponse[] | null;
  /** @nullable */
  committee?: CommitteeSlotResponse[] | null;
}